Refactor user optimisation and add collision optimisation
